Add files via upload
Hello Coders..!

Completed Task 5 of my CodSoft Internship!🎉

1.Create a Student class to represent individual students. Include attributes such as name, roll number, grade, and any other relevant details.
2.Implement a StudentManagementSystem class to manage the collection of students. Include
methods to add a student, remove a student, search for a student, and display all students.
3 . Design the user interface for the Student Management System. This can be a console-based
interface or a graphical user interface (GUI) using libraries like Swing or JavaFX.
4. Implement methods to read and write student data to a storage medium, such as a file or a database.
5. Allow users to interact with the Student Management System by providing options such as
adding a new student, editing an existing student's information, searching for a student, displaying all
students, and exiting the application.
6. Implement input validation to ensure that required fields are not left empty and that the student
data is in the correct format. 
Excited to continue learning and growing through this internship!